+++ src/steps/cloud-asset/index.ts
@@ -434,6 +434,7 @@ export function buildIamTargetRelationship({
   logger,
   parsedMember,
   projectId,
+  projectName,
   condition,
 }: {
   bindingEntity: BindingEntity;
@@ -441,6 +442,7 @@ export function buildIamTargetRelationship({
   parsedMember: ParsedIamMember;
   logger: IntegrationLogger;
   projectId?: string;
+  projectName?: string;
   condition?: cloudresourcemanager_v3.Schema$Expr;
 }): Relationship | undefined {
   if (principalEntity) {
@@ -450,6 +452,7 @@ export function buildIamTargetRelationship({
       to: principalEntity,
       properties: {
         projectId,
+        projectName,
         ...(condition && getConditionRelationshipProperties(condition)),
       },
     });
@@ -484,6 +487,7 @@ export function buildIamTargetRelationship({
           },
           properties: {
             projectId,
+            projectName,
             ...(condition && getConditionRelationshipProperties(condition)),
           },
         })
@@ -524,6 +528,7 @@ async function createAndAddConvienceMemberTargetRelationships(
             to: targetRoleEntitiy,
             properties: {
               projectId: bindingEntity.projectId,
+              projectName: bindingEntity.projectName,
               ...(condition && getConditionRelationshipProperties(condition)),
             },
           })
@@ -549,6 +554,7 @@ async function createAndAddConvienceMemberTargetRelationships(
             },
             properties: {
               projectId: bindingEntity.projectId,
+              projectName: bindingEntity.projectName,
               ...(condition && getConditionRelationshipProperties(condition)),
             },
           }),
@@ -603,6 +609,7 @@ export async function createPrincipalRelationships(
               parsedMember,
               logger,
               projectId: bindingEntity.projectId,
+              projectName: bindingEntity.projectName,
               condition,
             }),
           );