+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Fixed
+
+- `_key`s are now generated properly to allow for mapping `google_iam_binding`s
+  to `google_bigquery_table`s when making `google_iam_binding_allows_resource`
+  relationships.
+
 ## 2.6.0 - 2021-12-07
 
 ### Added