+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.6.1 - 2021-12-14
+
 ### Fixed
 
 - `_key`s are now generated properly to allow for mapping `google_iam_binding`s